data_processor/pyproject.toml
[project]
name = "data-processor"
version = "0.1.0"
description = "Standalone Spark ETL for data cleaning in WSL"
dependencies = [
    "pyspark>=3.5.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.1.0",
    "pypdf>=4.0.0",
]
requires-python = ">=3.10"

[tool.uv]
package = false

